put preprocessing earlier in Dockerfile
